+++ Source/DlgSystem/Private/DlgObject.cpp
@@ -8,6 +8,7 @@
 
 void UDlgObject::PostInitProperties()
 {
+	// We must always set the outer to be something that exists at runtime
 #if WITH_EDITOR
 	if (UEdGraphNode* GraphNode =  Cast<UEdGraphNode>(GetOuter()))
 	{